requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py35_0
bleach=1.5.0=py35_0
cffi=1.9.1=py35_0
click=6.6=py35_0
decorator=4.0.10=py35_0
entrypoints=0.2.2=py35_1
flask=0.12.1=py35_0
freetype=2.5.5=2
h5py=2.6.0=np111py35_2
hdf5=1.8.17=1
html5lib=0.999=py35_0
ipykernel=4.5.2=py35_0
ipython=5.3.0=py35_0
ipython_genutils=0.1.0=py35_0
itsdangerous=0.24=py35_0
jbig=2.1=0
jinja2=2.9.5=py35_0
jpeg=9b=0
jsonschema=2.5.1=py35_0
jupyter_client=5.0.0=py35_0
jupyter_core=4.3.0=py35_0
libpng=1.6.27=0
libtiff=4.0.6=3
markupsafe=0.23=py35_2
memory_profiler=0.43=py35_0
mistune=0.7.4=py35_0
mkl=11.3.3=0
nbconvert=5.1.1=py35_0
nbformat=4.3.0=py35_0
networkx=1.11=py35_0
nltk=3.2.2=py35_0
notebook=4.4.1=py35_0
numpy=1.11.2=py35_0
olefile=0.44=py35_0
openssl=1.0.2j=0
pandocfilters=1.4.1=py35_0
path.py=10.1=py35_0
pexpect=4.2.1=py35_0
pickleshare=0.7.4=py35_0
pillow=4.0.0=py35_1
pip=9.0.1=py35_0
prompt_toolkit=1.0.13=py35_0
psutil=5.2.1=py35_0
ptyprocess=0.5.1=py35_0
py=1.4.31=py35_0
pycparser=2.17=py35_0
pygments=2.2.0=py35_0
pytest=3.0.5=py35_0
python=3.5.2=0
python-dateutil=2.6.0=py35_0
pytorch=0.1.12=py35_2cu75
pyyaml=3.12=py35_0
pyzmq=16.0.2=py35_0
readline=6.2=2
scipy=0.18.1=np111py35_0
setuptools=27.2.0=py35_0
simplegeneric=0.8.1=py35_1
six=1.10.0=py35_0
sqlite=3.13.0=0
terminado=0.6=py35_0
testpath=0.3=py35_0
tk=8.5.18=0
torchvision=0.1.7=py35_1
tornado=4.4.2=py35_0
tqdm=4.8.4=py35_0
traitlets=4.3.2=py35_0
wcwidth=0.1.7=py35_0
werkzeug=0.12.1=py35_0
wheel=0.29.0=py35_0
xz=5.2.2=0
yaml=0.1.6=0
zlib=1.2.8=3
